feat(core): rename help to defineStackTemplate
