+++ packages/core/src/defineStackTemplate.ts
