+++ packages/core/src/index.ts
@@ -2,5 +2,5 @@ export * from './logger.js';
 export * from './BaseProvider.js';
 export * from './BaseConnector.js';
 export * from './types.js';
-export * from './helper.js';
+export * from './defineStackTemplate.js';
 export * from './version.js';
